refactor(shared): extract emit_notification into shared.notifications
Notification emission is shared support code for agent loops. Centralize it so user-facing load failures can be surfaced consistently.
