+++ src/owui_ext/shared/notifications.py
@@ -1,0 +1,27 @@
+"""Frontend toast notification helper shared by tool wrappers."""
+
+import logging
+from typing import Callable, Optional
+
+_notifications_log = logging.getLogger("owui_ext.shared.notifications")
+
+
+async def emit_notification(
+    event_emitter: Optional[Callable], *, level: str, content: str
+) -> None:
+    """Emit a frontend notification toast when the current chat supports it."""
+    if not callable(event_emitter):
+        return
+    if not isinstance(content, str) or not content.strip():
+        return
+    try:
+        await event_emitter(
+            {
+                "type": "notification",
+                "data": {"type": level, "content": content.strip()},
+            }
+        )
+    except Exception as exc:
+        _notifications_log.debug(
+            f"Error emitting notification ({level}): {exc}"
+        )
